:recycle: Move test code 'example_test.go' to 'pipeline_test.go'.
