+++ example_test.go
@@ -3,7 +3,6 @@ package pipeline
 import (
 	"fmt"
 	"log"
-	"testing"
 )
 
 func ExampleCommandPipeLine() {
@@ -18,28 +17,4 @@ func ExampleCommandPipeLine() {
 	fmt.Println(string(out))
 	// Output:
 	// 1
-}
-
-func TestCombinedOutput(t *testing.T) {
-	out, err := CombinedOutput(
-		[]string{"echo", "1"},
-		[]string{"grep", "2"},
-	)
-	if err == nil {
-		log.Fatal(err)
-	}
-	if string(out) != "" {
-		log.Fatal("output is not empty.")
-	}
-
-	out, err = CombinedOutput(
-		[]string{"echo", "1"},
-		[]string{"rmdir", "tmptmptmp"},
-	)
-	if err == nil {
-		log.Fatal(err)
-	}
-	if string(out) == "" {
-		log.Fatal("output is empty.")
-	}
 }
